requirements.txt
# Dependencias principales
google-adk>=0.1.0
pydantic>=2.0.0
google-generativeai>=0.3.0

# Utilidades adicionales
python-dotenv>=1.0.0
asyncio>=3.4.3
typing-extensions>=4.0.0
ratelimit>=2.2.1

# Para desarrollo y testing (opcional)
pytest>=7.0.0
pytest-asyncio>=0.21.0